feat: done: detect squash commits
Previously, if you did a GitHub Squash Merge, since the SHAs are
rewritten, `git wf done` would not detect that you could safely delete
the branch.  Now, if it doesn't find those exact commits, it examines
the commits since the merge-base with your parent branch and sees if any
of the diffs are identical to what it would expect.